+++ maps/uk/london.geo.json
@@ -1,0 +1,93 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "type": "Market",
+        "name": "London",
+        "name_long": "London",
+        "iso_a2": "London"
+      },
+      "geometry": {
+        "coordinates": [
+          [
+            [
+              0.4329942697808633,
+              51.471442744627296
+            ],
+            [
+              0.5739122349433501,
+              51.51729074554004
+            ],
+            [
+              0.7179732844541036,
+              51.52326669708225
+            ],
+            [
+              0.8940593946452964,
+              51.56905821671438
+            ],
+            [
+              0.9228718822305098,
+              51.64858412816602
+            ],
+            [
+              0.9004621696642516,
+              51.69226401199876
+            ],
+            [
+              0.7692033343302285,
+              51.72797044902347
+            ],
+            [
+              0.10923598690564518,
+              51.93160964660518
+            ],
+            [
+              -0.6493205488335718,
+              51.982905012907025
+            ],
+            [
+              -1.0558216122181818,
+              51.74571646324094
+            ],
+            [
+              -0.9277543944468789,
+              51.253559108947655
+            ],
+            [
+              -0.05727705645625747,
+              51.09301850353336
+            ],
+            [
+              0.7622470765952869,
+              51.23151614898231
+            ],
+            [
+              0.8490274418316801,
+              51.36162768837451
+            ],
+            [
+              0.6699720187166918,
+              51.39821451821396
+            ],
+            [
+              0.5995589328235553,
+              51.46806531105241
+            ],
+            [
+              0.5323271844178521,
+              51.48401529830565
+            ],
+            [
+              0.4329942697808633,
+              51.471442744627296
+            ]
+          ]
+        ],
+        "type": "Polygon"
+      }
+    }
+  ]
+}
