+++ maps/uk/midlands.geo.json
@@ -1,0 +1,229 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "type": "Market",
+        "name": "Midlands",
+        "name_long": "Midlands",
+        "iso_a2": "Midlands"
+      },
+      "geometry": {
+        "coordinates": [
+          [
+            [
+              -0.6352673032739915,
+              53.72557527397794
+            ],
+            [
+              -1.0573290603408907,
+              53.31733905760879
+            ],
+            [
+              -1.9573048640079946,
+              52.92007156832892
+            ],
+            [
+              -3.1619007823759375,
+              52.792082420849795
+            ],
+            [
+              -3.0067530888686065,
+              52.76394991318875
+            ],
+            [
+              -2.955073601893389,
+              52.716992401617176
+            ],
+            [
+              -3.084429834397639,
+              52.65742622154082
+            ],
+            [
+              -3.1154005075992757,
+              52.544397459395014
+            ],
+            [
+              -3.001868859583368,
+              52.582082071677036
+            ],
+            [
+              -3.0120521982834703,
+              52.50664941843519
+            ],
+            [
+              -3.1736585483484987,
+              52.503453861513634
+            ],
+            [
+              -3.230752138577742,
+              52.43383678881111
+            ],
+            [
+              -3.076400185407863,
+              52.3428053045871
+            ],
+            [
+              -2.963205337156438,
+              52.352229265862945
+            ],
+            [
+              -2.988764643116724,
+              52.27359883857335
+            ],
+            [
+              -3.091845111254031,
+              52.18226098242653
+            ],
+            [
+              -3.1173757835027516,
+              52.05897740308882
+            ],
+            [
+              -3.0064056069232095,
+              51.91615498417332
+            ],
+            [
+              -2.885315447336893,
+              51.91923882490701
+            ],
+            [
+              -2.661678653026172,
+              51.830263397843794
+            ],
+            [
+              -2.7329838965025317,
+              51.58547297684606
+            ],
+            [
+              -2.4236523893980007,
+              51.74159818191424
+            ],
+            [
+              -1.6485913112305752,
+              51.95012112052734
+            ],
+            [
+              -0.6307402254797125,
+              51.982271674820026
+            ],
+            [
+              0.09961259359752717,
+              51.93537800107404
+            ],
+            [
+              0.7523847921932258,
+              51.74001597141128
+            ],
+            [
+              0.9489972296913152,
+              51.808184591959645
+            ],
+            [
+              1.1447206424338958,
+              51.7816131033658
+            ],
+            [
+              1.2669100599120497,
+              51.83976298903795
+            ],
+            [
+              1.281047541778861,
+              51.917738002847074
+            ],
+            [
+              1.2302969864775264,
+              51.97452244535896
+            ],
+            [
+              1.3317395920038564,
+              51.963160617059145
+            ],
+            [
+              1.6084976343260848,
+              52.14724406020133
+            ],
+            [
+              1.6559557909900775,
+              52.30722361301119
+            ],
+            [
+              1.73856043122737,
+              52.472342702682624
+            ],
+            [
+              1.734128844265797,
+              52.650091446659815
+            ],
+            [
+              1.6617938464028157,
+              52.76096209443651
+            ],
+            [
+              1.5093382162316402,
+              52.84544818647362
+            ],
+            [
+              1.3014830500834762,
+              52.92893197833931
+            ],
+            [
+              0.9381304615910437,
+              52.95133799591281
+            ],
+            [
+              0.5540649639751507,
+              52.97123768199168
+            ],
+            [
+              0.4632021074450563,
+              52.87413594390728
+            ],
+            [
+              0.28574137729754057,
+              52.809295344292394
+            ],
+            [
+              0.06273146372339511,
+              52.9065439291835
+            ],
+            [
+              0.302260630154791,
+              53.0805464900491
+            ],
+            [
+              0.3642078283706667,
+              53.16233324282737
+            ],
+            [
+              0.28161156408322086,
+              53.34023732704591
+            ],
+            [
+              0.15358735443834348,
+              53.4436733021347
+            ],
+            [
+              -0.07357398847861418,
+              53.568928405468625
+            ],
+            [
+              -0.2924407014501753,
+              53.7060504760025
+            ],
+            [
+              -0.44958910621150494,
+              53.6912806138136
+            ],
+            [
+              -0.6352673032739915,
+              53.72557527397794
+            ]
+          ]
+        ],
+        "type": "Polygon"
+      }
+    }
+  ]
+}
