+++ maps/uk/north_east.geo.json
@@ -1,0 +1,110 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "type": "Market",
+        "name": "North East",
+        "name_long": "North East",
+        "iso_a2": "North East"
+      },
+      "geometry": {
+        "coordinates": [
+          [
+            [
+              -1.5696030763697593,
+              55.40514174192907
+            ],
+            [
+              -1.6481817141972215,
+              55.579998833358616
+            ],
+            [
+              -2.012162482074018,
+              55.7851984674171
+            ],
+            [
+              -2.2488152494333065,
+              55.64662433776084
+            ],
+            [
+              -2.3307466284755947,
+              55.64135496343974
+            ],
+            [
+              -2.1899558857299155,
+              55.463291144540364
+            ],
+            [
+              -2.3718206603235785,
+              55.3601409498182
+            ],
+            [
+              -2.6219459522258717,
+              55.25406041535879
+            ],
+            [
+              -2.8675310601732633,
+              55.101024050326316
+            ],
+            [
+              -2.9539572845613122,
+              55.051715169839355
+            ],
+            [
+              -3.063141625313648,
+              55.0517274461308
+            ],
+            [
+              -3.058603921021813,
+              54.96053040962437
+            ],
+            [
+              -3.2724478985853693,
+              54.90565326295183
+            ],
+            [
+              -3.468087155344648,
+              54.7746427908437
+            ],
+            [
+              -3.586319159625475,
+              54.545602643428424
+            ],
+            [
+              -3.577243277244804,
+              54.471674414976434
+            ],
+            [
+              -3.3678096341311345,
+              54.27801703003135
+            ],
+            [
+              -0.4746348918968124,
+              54.347184993156645
+            ],
+            [
+              -0.6611017028523065,
+              54.5007087172267
+            ],
+            [
+              -1.2022970601428824,
+              54.68268110476345
+            ],
+            [
+              -1.3524781423157322,
+              54.90822319928705
+            ],
+            [
+              -1.5696030763697593,
+              55.40514174192907
+            ]
+          ]
+        ],
+        "type": "Polygon"
+      },
+      "id": 1
+    }
+  ]
+}
