+++ maps/uk/north_west.geo.json
@@ -1,0 +1,201 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "type": "Market",
+        "name": "North West",
+        "name_long": "North West",
+        "iso_a2": "North West"
+      },
+      "geometry": {
+        "coordinates": [
+          [
+            [
+              -0.46885852617648993,
+              54.34617545107275
+            ],
+            [
+              -3.3608803366292364,
+              54.27457580424567
+            ],
+            [
+              -3.175705070824762,
+              54.1320011950059
+            ],
+            [
+              -3.1164532880731315,
+              54.129827764117664
+            ],
+            [
+              -2.986810851521483,
+              54.16891574337356
+            ],
+            [
+              -2.868318584471723,
+              54.175389887411285
+            ],
+            [
+              -2.849787837221271,
+              54.14071063194373
+            ],
+            [
+              -2.9090321789007305,
+              53.955904509587384
+            ],
+            [
+              -3.027546823673987,
+              53.914479007588426
+            ],
+            [
+              -3.0534738499650302,
+              53.820572006880866
+            ],
+            [
+              -3.03125460129678,
+              53.77244445751694
+            ],
+            [
+              -2.9275550581918424,
+              53.733028207458005
+            ],
+            [
+              -3.0016200416025356,
+              53.64750183799458
+            ],
+            [
+              -3.0608777512245524,
+              53.55959765585976
+            ],
+            [
+              -3.068290023784158,
+              53.50896925458758
+            ],
+            [
+              -3.0016389687047536,
+              53.41195093753623
+            ],
+            [
+              -2.909048476942445,
+              53.36114808687083
+            ],
+            [
+              -2.816452360254914,
+              53.33019290387449
+            ],
+            [
+              -2.749808630110323,
+              53.31471269737594
+            ],
+            [
+              -2.8646045186282265,
+              53.2970058400052
+            ],
+            [
+              -2.9127535693841367,
+              53.30364703833749
+            ],
+            [
+              -3.0608878881511714,
+              53.42298589677134
+            ],
+            [
+              -3.1645787430639984,
+              53.400908761991275
+            ],
+            [
+              -3.0831149428568096,
+              53.26157970493762
+            ],
+            [
+              -2.9090395700355884,
+              53.20615431843348
+            ],
+            [
+              -2.968308519938745,
+              53.15289449973372
+            ],
+            [
+              -2.8868461897519637,
+              53.08400239480093
+            ],
+            [
+              -2.8128060180818864,
+              52.990499556017
+            ],
+            [
+              -2.716533220464271,
+              52.96151853950832
+            ],
+            [
+              -2.7906128395117378,
+              52.892333109774626
+            ],
+            [
+              -2.849856144182638,
+              52.936986861117845
+            ],
+            [
+              -2.97576259040207,
+              52.95261140908735
+            ],
+            [
+              -3.105451781446675,
+              52.88357916252281
+            ],
+            [
+              -3.1720382860726772,
+              52.79174177565989
+            ],
+            [
+              -1.9797834740451208,
+              52.92141560651919
+            ],
+            [
+              -1.0499747150377345,
+              53.32135157972348
+            ],
+            [
+              -0.6498598431215044,
+              53.72426519163898
+            ],
+            [
+              -0.4464187388400944,
+              53.724243186906335
+            ],
+            [
+              -0.2757755432633928,
+              53.735222295389605
+            ],
+            [
+              -0.07206800796305401,
+              53.647501785318525
+            ],
+            [
+              0.035337101315889186,
+              53.645306438506964
+            ],
+            [
+              0.11311332042430422,
+              53.61675618371547
+            ],
+            [
+              -0.20545971373843486,
+              54.01034143760441
+            ],
+            [
+              -0.0757706289977591,
+              54.12119967642204
+            ],
+            [
+              -0.46885852617648993,
+              54.34617545107275
+            ]
+          ]
+        ],
+        "type": "Polygon"
+      }
+    }
+  ]
+}
