+++ maps/uk/south.geo.json
@@ -1,0 +1,90 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "type": "Market",
+        "name": "South",
+        "name_long": "South",
+        "iso_a2": "South"
+      },
+      "geometry": {
+        "coordinates": [
+          [
+            [-0.6620033914081773, 51.981500690565355],
+            [-1.6414698490734736, 51.95183445046172],
+            [-2.4415686373495475, 51.73446699129974],
+            [-2.7968967910196625, 51.46406830399803],
+            [-3.0305953562795764, 51.25619080551451],
+            [-3.1265897593897023, 51.208082546290655],
+            [-3.2898453835057637, 51.190030220686054],
+            [-3.5841652442452983, 51.22612284568979],
+            [-3.8335009882680424, 51.232127334796985],
+            [-4.169963922849064, 51.20206409940991],
+            [-4.288557796149831, 51.03328006677975],
+            [-4.531793639002331, 50.97285426144899],
+            [-4.576473742429442, 50.79111895583645],
+            [-4.86791237307915, 50.578142773857905],
+            [-4.893423064041627, 50.53546119592755],
+            [-4.973501768616671, 50.52323855976434],
+            [-5.281044395441597, 50.29059733875738],
+            [-5.344673894219028, 50.247750437278796],
+            [-5.5689606640585225, 50.19853135740928],
+            [-5.666423163512292, 50.126286846216146],
+            [-5.666185897354069, 50.07954064373388],
+            [-5.623251106808851, 50.04780502370977],
+            [-5.537820341140446, 50.085801695870856],
+            [-5.424701203337008, 50.1028479353954],
+            [-5.33470978546373, 50.08576542750569],
+            [-5.2348095515970385, 50.02371897175868],
+            [-5.1089178611200055, 50.041242992685625],
+            [-5.03383891605074, 50.135446359297674],
+            [-4.788020436770211, 50.26658390796254],
+            [-4.518033316525731, 50.33645830939406],
+            [-4.285786559498774, 50.35824813278512],
+            [-4.19072714700809, 50.38205565610801],
+            [-4.072266868447116, 50.33082807489117],
+            [-3.910245986872866, 50.29091794728615],
+            [-3.7851688209047154, 50.22783976874419],
+            [-3.672317352040949, 50.24033267726668],
+            [-3.5712877825305043, 50.332214943927596],
+            [-3.4941400487564067, 50.464178762861536],
+            [-3.4054895398388396, 50.620298010897955],
+            [-2.9729212411882884, 50.71606745130512],
+            [-2.7090708456955213, 50.68945852606046],
+            [-2.5399284328517524, 50.612360843173576],
+            [-2.4271082326062015, 50.595839140481104],
+            [-2.362039912967049, 50.63439771792568],
+            [-2.0183731481170923, 50.59896468518187],
+            [-1.9446225554169132, 50.629221679989314],
+            [-2.0009551485612462, 50.725490450009886],
+            [-1.8704651151606981, 50.70887861243128],
+            [-1.5885859605837993, 50.731175438975924],
+            [-1.3243271570741229, 50.81196586309622],
+            [-1.3859930253254618, 50.889836811081295],
+            [-1.14373569461506, 50.77296707006812],
+            [-0.9020103140964864, 50.77873358514083],
+            [-0.7822294435024162, 50.72626399029875],
+            [-0.3431669795002392, 50.80765593921356],
+            [-0.04923415700417877, 50.79240808198281],
+            [0.24767874486752817, 50.730813639622596],
+            [0.5845288368164745, 50.85113813022724],
+            [0.7919991609961698, 50.92281030576058],
+            [0.9849023691279797, 50.91034533035946],
+            [1.0450071270702779, 51.02718965677039],
+            [1.416028225083835, 51.16574038787613],
+            [1.4487859404540586, 51.37544128075825],
+            [0.8578532055024368, 51.3621588979845],
+            [0.7671817937178673, 51.23238980200472],
+            [-0.05682563053133549, 51.09210580887947],
+            [-0.934177780147337, 51.25025023236313],
+            [-1.0602990570349675, 51.74527904413654],
+            [-0.6620033914081773, 51.981500690565355]
+          ]
+        ],
+        "type": "Polygon"
+      }
+    }
+  ]
+}
