--- se_low.geo.json
@@ -1,1 +1,0 @@
-{"type":"FeatureCollection","features":[{"type":"Feature","properties":{"featurecla":"Admin-0 country","scalerank":1,"labelrank":3,"sovereignt":"Sweden","sov_a3":"SWE","adm0_dif":0,"level":2,"type":"Sovereign country","tlc":"1","admin":"Sweden","adm0_a3":"SWE","geou_dif":0,"geounit":"Sweden","gu_a3":"SWE","su_dif":0,"subunit":"Sweden","su_a3":"SWE","brk_diff":0,"name":"Sweden","name_long":"Sweden","brk_a3":"SWE","brk_name":"Sweden","brk_group":null,"abbrev":"Swe.","postal":"S","formal_en":"Kingdom of Sweden","formal_fr":null,"name_ciawf":"Sweden","note_adm0":null,"note_brk":null,"name_sort":"Sweden","name_alt":null,"mapcolor7":1,"mapcolor8":4,"mapcolor9":2,"mapcolor13":4,"pop_est":10285453,"pop_rank":14,"pop_year":2019,"gdp_md":530883,"gdp_year":2019,"economy":"2. Developed region: nonG7","income_grp":"1. High income: OECD","fips_10":"SW","iso_a2":"SE","iso_a2_eh":"SE","iso_a3":"SWE","iso_a3_eh":"SWE","iso_n3":"752","iso_n3_eh":"752","un_a3":"752","wb_a2":"SE","wb_a3":"SWE","woe_id":23424954,"woe_id_eh":23424954,"woe_note":"Exact WOE match as country","adm0_iso":"SWE","adm0_diff":null,"adm0_tlc":"SWE","adm0_a3_us":"SWE","adm0_a3_fr":"SWE","adm0_a3_ru":"SWE","adm0_a3_es":"SWE","adm0_a3_cn":"SWE","adm0_a3_tw":"SWE","adm0_a3_in":"SWE","adm0_a3_np":"SWE","adm0_a3_pk":"SWE","adm0_a3_de":"SWE","adm0_a3_gb":"SWE","adm0_a3_br":"SWE","adm0_a3_il":"SWE","adm0_a3_ps":"SWE","adm0_a3_sa":"SWE","adm0_a3_eg":"SWE","adm0_a3_ma":"SWE","adm0_a3_pt":"SWE","adm0_a3_ar":"SWE","adm0_a3_jp":"SWE","adm0_a3_ko":"SWE","adm0_a3_vn":"SWE","adm0_a3_tr":"SWE","adm0_a3_id":"SWE","adm0_a3_pl":"SWE","adm0_a3_gr":"SWE","adm0_a3_it":"SWE","adm0_a3_nl":"SWE","adm0_a3_se":"SWE","adm0_a3_bd":"SWE","adm0_a3_ua":"SWE","adm0_a3_un":-99,"adm0_a3_wb":-99,"continent":"Europe","region_un":"Europe","subregion":"Northern Europe","region_wb":"Europe & Central Asia","name_len":6,"long_len":6,"abbrev_len":4,"tiny":-99,"homepart":1,"min_zoom":0,"min_label":2,"max_label":7,"label_x":19.01705,"label_y":65.85918,"ne_id":1159321287,"wikidataid":"Q34","name_ar":"السويد","name_bn":"সুইডেন","name_de":"Schweden","name_en":"Sweden","name_es":"Suecia","name_fa":"سوئد","name_fr":"Suède","name_el":"Σουηδία","name_he":"שוודיה","name_hi":"स्वीडन","name_hu":"Svédország","name_id":"Swedia","name_it":"Svezia","name_ja":"スウェーデン","name_ko":"스웨덴","name_nl":"Zweden","name_pl":"Szwecja","name_pt":"Suécia","name_ru":"Швеция","name_sv":"Sverige","name_tr":"İsveç","name_uk":"Швеція","name_ur":"سویڈن","name_vi":"Thụy Điển","name_zh":"瑞典","name_zht":"瑞典","fclass_iso":"Admin-0 country","tlc_diff":null,"fclass_tlc":"Admin-0 country","fclass_us":null,"fclass_fr":null,"fclass_ru":null,"fclass_es":null,"fclass_cn":null,"fclass_tw":null,"fclass_in":null,"fclass_np":null,"fclass_pk":null,"fclass_de":null,"fclass_gb":null,"fclass_br":null,"fclass_il":null,"fclass_ps":null,"fclass_sa":null,"fclass_eg":null,"fclass_ma":null,"fclass_pt":null,"fclass_ar":null,"fclass_jp":null,"fclass_ko":null,"fclass_vn":null,"fclass_tr":null,"fclass_id":null,"fclass_pl":null,"fclass_gr":null,"fclass_it":null,"fclass_nl":null,"fclass_se":null,"fclass_bd":null,"fclass_ua":null,"filename":"SWE.geojson"},"geometry":{"type":"Polygon","coordinates":[[[11.027368605196868,58.85614940045936],[11.468271925511146,59.43239329694604],[12.3003658382749,60.11793284773003],[12.631146681375185,61.293571682370136],[11.992064243221563,61.80036245385655],[11.930569288794231,63.12831757267698],[12.579935336973934,64.06621898055833],[13.571916131248713,64.04911408146971],[13.919905226302204,64.44542064071608],[13.55568973150909,64.78702769638151],[15.108411492583002,66.19386688909547],[16.108712192456778,67.30245555283689],[16.768878614985482,68.0139366726314],[17.729181756265348,68.01055186631628],[17.993868442464333,68.56739126247736],[19.878559604581255,68.40719432237258],[20.025268995857886,69.0651386583127],[20.645592889089528,69.10624726020087],[21.978534783626117,68.6168456081807],[23.53947309743444,67.93600861273525],[23.565879754335583,66.39605093043743],[23.903378533633802,66.00692739527962],[22.18317345550193,65.72374054632017],[21.21351687997722,65.02600535751527],[21.369631381930958,64.41358795842429],[19.77887576669022,63.60955434839504],[17.84777916837521,62.74940013289681],[17.119554884518124,61.34116567651097],[17.83134606290639,60.63658336042741],[18.78772179533209,60.081914374422595],[17.86922488777634,58.9537661810587],[16.829185011470088,58.71982697207339],[16.447709588291474,57.041118069071885],[15.879785597403783,56.10430186626866],[14.666681349352075,56.200885118222175],[14.100721062891465,55.40778107362265],[12.942910597392057,55.36173737245058],[12.625100538797028,56.30708018658197],[11.787942335668674,57.44181712506307],[11.027368605196868,58.85614940045936]]]}}]}
