--- uk_low.geo.json
@@ -1,1 +1,0 @@
-{"type":"FeatureCollection","features":[{"type":"Feature","properties":{"featurecla":"Admin-0 country","scalerank":1,"labelrank":2,"sovereignt":"United Kingdom","sov_a3":"GB1","adm0_dif":1,"level":2,"type":"Country","tlc":"1","admin":"United Kingdom","adm0_a3":"GBR","geou_dif":0,"geounit":"United Kingdom","gu_a3":"GBR","su_dif":0,"subunit":"United Kingdom","su_a3":"GBR","brk_diff":0,"name":"United Kingdom","name_long":"United Kingdom","brk_a3":"GBR","brk_name":"United Kingdom","brk_group":null,"abbrev":"U.K.","postal":"GB","formal_en":"United Kingdom of Great Britain and Northern Ireland","formal_fr":null,"name_ciawf":"United Kingdom","note_adm0":null,"note_brk":null,"name_sort":"United Kingdom","name_alt":null,"mapcolor7":6,"mapcolor8":6,"mapcolor9":6,"mapcolor13":3,"pop_est":66834405,"pop_rank":16,"pop_year":2019,"gdp_md":2829108,"gdp_year":2019,"economy":"1. Developed region: G7","income_grp":"1. High income: OECD","fips_10":"UK","iso_a2":"GB","iso_a2_eh":"GB","iso_a3":"GBR","iso_a3_eh":"GBR","iso_n3":"826","iso_n3_eh":"826","un_a3":"826","wb_a2":"GB","wb_a3":"GBR","woe_id":-90,"woe_id_eh":23424975,"woe_note":"Eh ID includes Channel Islands and Isle of Man. UK constituent countries of England (24554868), Wales (12578049), Scotland (12578048), and Northern Ireland (20070563).","adm0_iso":"GBR","adm0_diff":null,"adm0_tlc":"GBR","adm0_a3_us":"GBR","adm0_a3_fr":"GBR","adm0_a3_ru":"GBR","adm0_a3_es":"GBR","adm0_a3_cn":"GBR","adm0_a3_tw":"GBR","adm0_a3_in":"GBR","adm0_a3_np":"GBR","adm0_a3_pk":"GBR","adm0_a3_de":"GBR","adm0_a3_gb":"GBR","adm0_a3_br":"GBR","adm0_a3_il":"GBR","adm0_a3_ps":"GBR","adm0_a3_sa":"GBR","adm0_a3_eg":"GBR","adm0_a3_ma":"GBR","adm0_a3_pt":"GBR","adm0_a3_ar":"GBR","adm0_a3_jp":"GBR","adm0_a3_ko":"GBR","adm0_a3_vn":"GBR","adm0_a3_tr":"GBR","adm0_a3_id":"GBR","adm0_a3_pl":"GBR","adm0_a3_gr":"GBR","adm0_a3_it":"GBR","adm0_a3_nl":"GBR","adm0_a3_se":"GBR","adm0_a3_bd":"GBR","adm0_a3_ua":"GBR","adm0_a3_un":-99,"adm0_a3_wb":-99,"continent":"Europe","region_un":"Europe","subregion":"Northern Europe","region_wb":"Europe & Central Asia","name_len":14,"long_len":14,"abbrev_len":4,"tiny":-99,"homepart":1,"min_zoom":0,"min_label":1.7,"max_label":6.7,"label_x":-2.116346,"label_y":54.402739,"ne_id":1159320713,"wikidataid":"Q145","name_ar":"المملكة المتحدة","name_bn":"যুক্তরাজ্য","name_de":"Vereinigtes Königreich","name_en":"United Kingdom","name_es":"Reino Unido","name_fa":"بریتانیا","name_fr":"Royaume-Uni","name_el":"Ηνωμένο Βασίλειο","name_he":"הממלכה המאוחדת","name_hi":"यूनाइटेड किंगडम","name_hu":"Egyesült Királyság","name_id":"Britania Raya","name_it":"Regno Unito","name_ja":"イギリス","name_ko":"영국","name_nl":"Verenigd Koninkrijk","name_pl":"Wielka Brytania","name_pt":"Reino Unido","name_ru":"Великобритания","name_sv":"Storbritannien","name_tr":"Birleşik Krallık","name_uk":"Велика Британія","name_ur":"مملکت متحدہ","name_vi":"Vương quốc Liên hiệp Anh và Bắc Ireland","name_zh":"英国","name_zht":"英國","fclass_iso":"Admin-0 country","tlc_diff":null,"fclass_tlc":"Admin-0 country","fclass_us":null,"fclass_fr":null,"fclass_ru":null,"fclass_es":null,"fclass_cn":null,"fclass_tw":null,"fclass_in":null,"fclass_np":null,"fclass_pk":null,"fclass_de":null,"fclass_gb":null,"fclass_br":null,"fclass_il":null,"fclass_ps":null,"fclass_sa":null,"fclass_eg":null,"fclass_ma":null,"fclass_pt":null,"fclass_ar":null,"fclass_jp":null,"fclass_ko":null,"fclass_vn":null,"fclass_tr":null,"fclass_id":null,"fclass_pl":null,"fclass_gr":null,"fclass_it":null,"fclass_nl":null,"fclass_se":null,"fclass_bd":null,"fclass_ua":null,"filename":"GBR.geojson"},"geometry":{"type":"MultiPolygon","coordinates":[[[[-6.197884894220991,53.867565009163364],[-6.953730231138067,54.073702297575636],[-7.572167934591064,54.059956366586],[-7.366030646178785,54.59584096945272],[-7.572167934591064,55.13162221945487],[-6.733847011736145,55.17286001242378],[-5.661948614921968,54.55460317648381],[-6.197884894220991,53.867565009163364]]],[[[-3.093830673788716,53.404547400669685],[-3.092079637047107,53.40444082296358],[-2.945008510744344,53.98499970154671],[-3.614700825433033,54.600936773292574],[-3.630005458989331,54.615012925833014],[-4.844169073903061,54.790971177786844],[-5.082526617849283,55.06160065369937],[-4.7191121077567,55.50847260194348],[-5.047980922862109,55.78398550070756],[-5.586397670911197,55.31114614523682],[-5.644998745130238,56.27501496034483],[-6.149980841486411,56.785009670633485],[-5.786824713555291,57.81884837506465],[-5.009998745127632,58.63001333275008],[-4.211494513353557,58.55084503847911],[-3.005004848635281,58.63500010846633],[-4.073828497728073,57.55302480735526],[-3.055001796877718,57.69001902936094],[-1.959280564776918,57.68479970969952],[-2.219988165689358,56.87001740175356],[-3.119003058271176,55.973793036515474],[-2.085009324543023,55.9099984808513],[-2.005675679673857,55.804902850350175],[-1.11499101399221,54.624986477265395],[-0.4304849918542,54.46437612570219],[0.184981316742039,53.32501414653103],[0.469976840831805,52.929999498092],[1.681530795914682,52.739520168664],[1.559987827164321,52.09999848083601],[1.050561557630942,51.806760565795685],[1.449865349950244,51.28942780212191],[0.550333693045502,50.765738837275876],[-0.787517462558696,50.77498891865622],[-2.489997524414434,50.50001862243124],[-2.956273972984093,50.696879991247044],[-3.617448085942385,50.22835561787275],[-4.542507900399244,50.341837063185665],[-5.245023159191135,49.95999990498109],[-5.776566941745273,50.15967763935686],[-4.309989793301895,51.21000112568919],[-3.414850633142123,51.42600861266925],[-3.422719467108379,51.42684816740609],[-4.98436723471093,51.593466091511004],[-5.267295701508942,51.991400458374585],[-4.222346564134909,52.301355699261364],[-4.77001339356417,52.840004991255626],[-4.579999152026971,53.49500377055517],[-3.093830673788716,53.404547400669685]]]]}}]}
